Create README.md
Am trying to get my librelink to connect to xdrip for carb bolus calc, open loop initially.
Get as far as ./ which command is not recognised. German site suggests win current apktool needs replacing with version 2.3.
which when downloaded  is apktool_2.3.0.jar cannot seem to load this into github or debian- would be most grateful if anyone can help?